chore: Rename handler class, move user helper
Moves the user helper function to the User file,
renames requesthandler.
